Attempt at scoring based on user input
This commit adds two new options.  The first will create a file
that lists all the services for the user to rank depending on how
critical they are for their environment.  The second option uses
the user scores to generate criticality of individual permissions
based on whether it is mutating and how critical the service is.

The algorithm is super basic right now but could be extended and
customized to use other fields.